server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
pydantic>=2.0.0
